Pass KUBERNETES_VERSION to MasterConfiguration
Previously KUBERNETES_VERSION was actually unused because it's not exported
with "export" keyword for "kubeadm" to pick it up. No matter what version was
given, it was reverting back to what's specified in the MasterConfiguration.

Even though I exported this variable to kubeadm (and directly passed to kubadm
with --kubernetes-version), it was still picking up what's in MasterConfiguration.

With this change, the specified version number is picked up according to
"kubectl version" and logs from kubeadm:

    INFO startup-script-url: [init] Using Kubernetes version: v1.8.0-beta.1

Bonus: Adding "set -e" for robustness and "set -x" for log verbosity.

Signed-off-by: Ahmet Alp Balkan <ahmetb@google.com>